Interesses cliente

+++ src/main/java/br/ufrn/alugai/repository/VendedorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.ufrn.alugai.model.Vendedor;
 
-public interface VendedorRepository extends JpaRepository<Vendedor, Integer>{
+public interface VendedorRepository extends JpaRepository<Vendedor, Long>{
 
 }
